migrate usage of old merkletree to the new merkletree impl in POD2 code
