+++ src/constants.rs
@@ -1,0 +1,1 @@
+pub const MAX_DEPTH: usize = 32;
