+++ src/primitives/mod.rs
@@ -1,2 +1,1 @@
 pub mod merkletree;
-pub mod merkletree_new;
